# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Steel Fabrication MRP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimates]
      summary: List estimates
      description: List all estimates, optionally filtered by status or search term.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, submitted, won, lost]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A list of estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Estimate"
    post:
      operationId: createEstimate
      tags: [estimates]
      summary: Create an estimate
      description: Create an estimate. The bid number is auto-assigned by the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /estimates/{estimateId}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      description: Update estimate fields or status (e.g. mark Lost). Won status is set via the convert endpoint.
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      description: Delete an estimate. Converted (won) estimates cannot be deleted.
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Cannot delete a converted estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /estimates/{estimateId}/convert:
    post:
      operationId: convertEstimateToJob
      tags: [estimates]
      summary: Mark estimate Won and convert it to a job
      description: >-
        Marks the estimate Won and creates a job linked to it with an
        auto-assigned job number and the given (or default) stage routing.
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateConvertInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Estimate already converted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /employees/{employeeId}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: integer
...
    EstimateId:
      name: estimateId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    JobDetail:
      type: object
      required:
        - id
        - jobNumber
        - name
        - customer
        - status
        - createdAt
        - updatedAt
        - stages
        - estimatedHours
        - actualHours
        - percentComplete
        - isPastDue
        - assignedEmployees
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        name:
          type: string
        customer:
          type: string
        customerId:
          type: ["integer", "null"]
        customerPo:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, complete, closed]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentStageName:
          type: ["string", "null"]
        estimatedHours:
          type: number
        actualHours:
          type: number
        percentComplete:
          type: number
        isPastDue:
          type: boolean
        estimateId:
          type: ["integer", "null"]
        bidNumber:
          type: ["string", "null"]
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
        assignedEmployees:
          type: array
          items:
            $ref: "#/components/schemas/AssignedEmployee"
        bomAssemblies:
          type: array
          items:
            $ref: "#/components/schemas/BomAssembly"
        assemblyCount:
          type: integer
        assemblyProgressPct:
          type: [number, "null"]
        assemblyStatus:
          type: [string, "null"]
    JobStageInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        estimatedHours:
          type: number
    Estimate:
      type: object
      required:
        - id
        - bidNumber
        - name
        - customer
        - status
        - type
        - marginPercent
        - quoteFormat
        - estimatedHours
        - jobId
        - jobNumber
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        bidNumber:
          type: string
        name:
          type: string
        customer:
          type: string
        status:
          type: string
          enum: [draft, submitted, won, lost]
        type:
          type: string
          enum: [preliminary, detailed]
          description: >-
            Internal values only — user-facing labels come from a swappable
            display constant (see OPEN_QUESTIONS.md).
        marginPercent:
          type: number
          minimum: 0
          maximum: 100
        quoteFormat:
          type: string
          enum: [itemized, summary]
        estimatedHours:
          type: number
        amount:
          type: ["number", "null"]
        bidDate:
          type: ["string", "null"]
          description: Calendar date (YYYY-MM-DD)
        dueDate:
          type: ["string", "null"]
          description: Calendar date (YYYY-MM-DD)
        notes:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
          description: Linked job id when the estimate was converted
        jobNumber:
          type: ["string", "null"]
          description: Linked job number when the estimate was converted
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EstimateInput:
      type: object
      required: [name, customer]
      properties:
        name:
          type: string
          minLength: 1
        customer:
          type: string
          minLength: 1
        status:
          type: string
          enum: [draft, submitted, lost]
        type:
          type: string
          enum: [preliminary, detailed]
        marginPercent:
          type: number
          minimum: 0
          maximum: 100
        quoteFormat:
          type: string
          enum: [itemized, summary]
        estimatedHours:
          type: number
        amount:
          type: ["number", "null"]
        bidDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EstimateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        customer:
          type: string
          minLength: 1
        status:
          type: string
          enum: [draft, submitted, lost]
        type:
          type: string
          enum: [preliminary, detailed]
        marginPercent:
          type: number
          minimum: 0
          maximum: 100
        quoteFormat:
          type: string
          enum: [itemized, summary]
        estimatedHours:
          type: number
        amount:
          type: ["number", "null"]
        bidDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EstimateConvertInput:
      type: object
      properties:
        dueDate:
          type: ["string", "null"]
        stages:
          type: array
          description: Custom routing. If omitted, the default workflow is applied.
          items:
            $ref: "#/components/schemas/JobStageInput"
...
    AssignedEmployee:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    Stage:
      type: object
      required:
        - id
        - jobId
        - name
        - orderIndex
        - status
        - estimatedHours
        - actualHours
        - createdAt
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        status:
          type: string
          enum: [not_started, in_progress, complete]
        estimatedHours:
          type: number
        actualHours:
          type: number
        createdAt:
          type: string
          format: date-time
...
    Employee:
      type: object
      required: [id, name, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        employeeCode:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        active:
          type: boolean
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        employeeCode:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        active:
          type: boolean
...
    BomPart:
      type: object
      required: [quantity]
      properties:
        id:
          type: integer
        partMark:
          type: [string, "null"]
        quantity:
          type: integer
          minimum: 1
        profileType:
          type: [string, "null"]
        profileSize:
          type: [string, "null"]
        grade:
          type: [string, "null"]
        lengthIn:
          type: [number, "null"]
        description:
          type: [string, "null"]
        heatNumber:
          type: [string, "null"]
    BomAssembly:
      type: object
      required: [mark, quantity, parts, onHold]
      properties:
        id:
          type: integer
        mark:
          type: string
        quantity:
          type: integer
          minimum: 1
        description:
          type: [string, "null"]
        finish:
          type: [string, "null"]
        processingPath:
          type: [string, "null"]
        currentStage:
          type: [string, "null"]
        onHold:
          type: boolean
        notes:
          type: [string, "null"]
        inspectedOn:
          type: [string, "null"]
        station:
          type: [string, "null"]
        inspector:
          type: [string, "null"]
        parts:
          type: array
          items:
            $ref: "#/components/schemas/BomPart"
...